#52
- error-tracking: раздел удален, содержимое перенесено в log (major)
- examples/server: мелкие правки стиля кода (patch)
