+++ src/log/index.ts
@@ -1,2 +1,12 @@
-export { Logger, LogEvent, LogLevel, LogHandler } from './types';
+export {
+  Logger,
+  LogEvent,
+  LogLevel,
+  LogHandler,
+  ErrorDetails,
+  BreadcrumbDetails,
+  ContextData,
+} from './types';
+
 export { createLogger } from './logger';
+export { DetailedError, Breadcrumb } from './errors';
