+++ src/preset/parts/__test__/utils.test.ts
@@ -5,8 +5,7 @@ import {
   AxiosResponse,
   InternalAxiosRequestConfig,
 } from 'axios';
-import { Breadcrumb, DetailedError } from '../../../error-tracking';
-import { Logger } from '../../../log';
+import { Logger, Breadcrumb, DetailedError } from '../../../log';
 import {
   HttpApiHostPool,
   HttpClientLogging,
